LayoutTests/fast/scrolling/ios/scrollto-does-not-interrupt-momentum-scroll-overflow.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true AsyncOverflowScrollingEnabled=true AsyncFrameScrollingEnabled=true CSSScrollAnchoringEnabled=false ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<style>
html, body {
    width: 100%;
    height: 100%;
    margin: 0px;
    padding: 0px;
}

#scroller {
    width: 100%;
    height: 400px;
    overflow: scroll;
    border: 1px solid black;
}

#content {
    width: 100px;
    height: 10000px;
    background: repeating-linear-gradient(to bottom, transparent, silver 200px);
}
</style>
<body>
    <div id="scroller">
        <div id="content"></div>
    </div>
    <p id="description"></p>
    <p id="console"></p>
</body>
<script>
jsTestIsAsync = true;

const scroller = document.getElementById("scroller");
let scrollTopBeforeScrollTo;

addEventListener("load", async () => {
    description("Verifies that calling scrollTo() during a momentum scroll does not interrupt the momentum scroll in an overflow scrolling container.");

    if (!window.testRunner)
        return;

    await UIHelper.dragFromPointToPoint(160, 350, 160, 0, 0.1);

    await shouldBecomeEqual("scroller.scrollTop > 300", "true");

    scrollTopBeforeScrollTo = scroller.scrollTop;
    scroller.scrollTo(0, 0);

    await UIHelper.waitForTargetScrollAnimationToSettle(scroller);

    shouldBeTrue("scroller.scrollTop > scrollTopBeforeScrollTo");

    finishJSTest();
});
</script>
</html>
