LayoutTests/fast/scrolling/ios/scrollto-does-not-interrupt-momentum-scroll-overflow-expected.txt

Verifies that calling scrollTo() during a momentum scroll does not interrupt the momentum scroll in an overflow scrolling container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollTop > 300 became true
PASS scroller.scrollTop > scrollTopBeforeScrollTo is true
PASS successfullyParsed is true

TEST COMPLETE
